@@ -1,39 +1,37 @@
 name: screenshots
 description: Auto-generation of screenshots for Apple and Play Stores using emulators, simulators and real devices. Includes support for multiple locales and framing. Compatible with fastlane.
 version: 2.1.1
-homepage: https://github.com/mmcc007/screenshots
-author: Maurice McCabe <mmcc007@gmail.com>
+homepage: https://github.com/pebble-world/screenshots
 
 environment:
-  sdk: '>=2.2.2 <3.0.0'
+  sdk: ">=2.16.1 <3.1.0"
 
 dependencies:
-  args: ^1.5.1
-  yaml: ^2.1.15
-  resource: ^2.1.5
-  path: ^1.6.2
-  file: ^5.0.7
-  archive: ^2.0.9
-  platform: ^2.2.0
-  process: ^3.0.9
-  meta: ^1.1.6
-  intl: ">=0.15.0 <1.0.0"
+  args: ^2.3.1
+  yaml: ^3.1.1
+  resource: ^2.1.7
+  path: ^1.8.2
+  file: ^6.1.2
+  archive: ^3.3.0
+  platform: ^3.1.0
+  process: ^4.2.4
+  meta: ^1.8.0
+  intl: ">=0.17.0 <1.0.0"
   tool_mobile: ^1.9.5
 #  tool_mobile:
 #    path: ../tool_mobile
 
 dev_dependencies:
-  test: ^1.5.1+1
-  test_api: ^0.2.5
-  mockito: ^4.1.0
-  pedantic: ^1.8.0+1
-  quiver: '>=2.0.0 <3.0.0'
-  fake_process_manager: ^0.1.0
+#  test: ^1.21.3
+#  test_api: ^0.4.11
+#  mockito: ^5.2.0
+#  quiver: ^3.1.0
+#  fake_process_manager: ^0.1.0
 #    path: ../fake_process_manager
-  tool_base_test:
-    git: https://github.com/mmcc007/tool_base_test.git
+#  tool_base_test:
+#    git: https://github.com/mmcc007/tool_base_test.git
 #    path: ../tool_base_test
-  collection: any
+#  collection: any
 
 executables:
   screenshots: main
